+++ setup_scripts/install-docker-debian.sh
@@ -42,6 +42,39 @@ sudo systemctl enable docker
 sudo apt-get -y install python-pip 
 sudo pip install --upgrade pip 
 sudo pip install netaddr 
+sudo apt-get -y install openssh-server
+
+#---Checking if packages have been installed. If not, the system will not reboot and allow the user to investigate.
+declare -a packagelist=("apt-transport-https" "ca-certificates" "curl" "gnupg2" "software-properties-common"  "docker-ce" "python-pip" "openssh-server")
+packagefail="false"
+
+for i in "${packagelist[@]}"
+do
+#echo $i
+packagecheck=$(dpkg -s $i 2> /dev/null | grep Status)
+#echo $packagecheck
+    if [ "$packagecheck" != "Status: install ok installed" ]; then
+       if [ $i = docker-ce ];then 
+           echo "ERROR: '$i' package did not install properly. Please check the terminal output above for any errors related to the pacakge installation. Run the install script two more times. If the issue persists, go to docker docs and follow the instructions for installing docker. (Make sure the instructions is CE and is for your Linux distribution,e.g., Ubuntu and Fedora.)"
+       else
+           echo "ERROR: '$i' package did not install properly. Please check the terminal output above for any errors related to the pacakge installation. Try installing the '$i' package individually by executing this in the command line: 'sudo apt-get install $i" 
+       fi
+       packagefail="true"
+       #echo $packagefail
+    fi
+done
+
+pipcheck=$(pip list 2> /dev/null | grep -F netaddr)
+#echo $pipcheck
+if [ -z "$pipcheck" ]; then
+    echo "ERROR: 'netaddr' package did not install properly. Please check the terminal output for any errors related to the pacakge installation. Make sure 'python-pip' is installed and then try running this command: 'sudo -H pip install netaddr' "
+    packagefail="true"
+    #echo $packagefail
+fi
+
+if [ $packagefail = "true" ]; then
+    exit
+fi
 
 sudo reboot
